tests for compareTwoObjects function was created.
