--- m1/bloco_10/testes-automatizados-com-jest_1/compareTwoObjects.js
@@ -1,19 +1,0 @@
-const assert = require('assert');
-
-const obj1 = {
-  title: 'My Title',
-  description: 'My Description',
-};
-
-const obj2 = {
-  description: 'My Description',
-  title: 'My Title',
-};
-
-const obj3 = {
-  title: 'My Different Title',
-  description: 'My Description',
-};
-
-// implemente seus testes aqui
-// Compare dois objetos (JSON) para verificar se são idênticos ou não
